requirements.txt
<<<<<<< HEAD
# Core
streamlit
google-generativeai

# Video / OpenCV
opencv-python-headless>=3.4.2
mediapipe>=0.8.6
streamlit-webrtc~=0.45.0

# Audio / Voz
twilio~=8.1.0
gTTS>=2.2.3
pygame==2.5.2
pydub==0.25.1
python-vlc==3.0.20123
playsound==1.3.0
=======
# Core
streamlit
google-generativeai

# Video / OpenCV
opencv-python-headless>=3.4.2
mediapipe>=0.8.6
streamlit-webrtc~=0.45.0

# Audio / Voz
twilio~=8.1.0
gTTS>=2.2.3
pygame==2.5.2
pydub==0.25.1
python-vlc==3.0.20123
playsound==1.3.0
>>>>>>> 91a58af7dfa2f97fb9f88daa8baa95310bfabacd
